+++ CotEditor/Sources/AlphaWindow.swift
@@ -94,6 +94,23 @@ final class AlphaWindow: NSWindow {
     }
     
     
+    /// apply current state to menu items
+    override func validateMenuItem(_ menuItem: NSMenuItem) -> Bool {
+        
+        // manually update the Japanese menu item title for toolbar visibility toggle
+        // since it doesn't work on macOS 10.12 and earlier (2018-05).
+        if floor(NSAppKitVersion.current.rawValue) <= NSAppKitVersion.macOS10_12.rawValue,
+            menuItem.action == #selector(toggleToolbarShown),
+            Locale.preferredLanguages.first == "ja",
+            let toolbar = self.toolbar
+        {
+            menuItem.title = toolbar.isVisible ? "ツールバーを非表示" : "ツールバーを表示"
+        }
+        
+        return super.validateMenuItem(menuItem)
+    }
+    
+    
     
     // MARK: Notifications